backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for the Rahasia encrypted sharing app"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.0",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.10.1",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
